# openapi: '3.0 api-docs.yml
openapi: "3.0.2"
info:
  title: Geotag Mobile Apps API Documentations
  description: this API just a references how the system on our apps is work, there might be some different Responses between the docs and Apps
  version: "BETA"
  contact:
    name: C23-GT02 (Ukasyah Nata)
    email: c128bsy3500@bangkit.academy
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: http://localhost:3000/api
paths:
  /auth/login:
    post:
      tags:
        - Authentication System
      summary: Sign-In to the Mobile Apps
      operationId: Sign In
      requestBody:
        description: Login into the application, will return a cookies session and roles on web browser, and return response data on mobile.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Login"
      responses:
        "200":
          description: OK, set cookies with user data
        "401":
          description: Authentication failed

  /auth/register:
    post:
      tags:
        - Authentication System
      summary: Register a new user for Mobile User
      operationId: Register Mobile
      requestBody:
        description: Create a new user account
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/Register"
      responses:
        "201":
          description: Created, redirected to login.
        "400":
          description: Bad Request
        "409":
          description: User already exists
        "500":
          description: Internal Server Error

  /user:
    get:
      tags:
        - User Profiles
      summary: Retrieve user data by email
      parameters:
        - in: query
          name: email
          required: true
          description: The email address of the user
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Get User Response'
        "404":
          description: User not found
          content:
            application/json:
              example:
                error: "User not found"

  /user/edit:
    post:
      tags:
        - User Profiles
      summary: Edit Information about User Authentication
      operationId: EditUser
      requestBody:
        description: Edit Existing User, image must be posted using html forms.
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/Edit User"
      responses:
        "200":
          description: Edited Succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Edit User Response"
        "400":
          description: Bad Request
        "413":
          description: Payload Too Large
        "500":
          description: Internal Server Error

  /user/reset:
    post:
      tags:
        - User Profiles
      summary: Reset User Password with generated links
      operationId: reserUser
      requestBody:
        description: Edit Existing User, image must be posted using html forms.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Reset Password"
      responses:
        "200":
          description: Reset Succesfully
        "500":
          description: Internal Server Error

components:
  schemas:
    Login:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minimum: 6
          format: password

    Register:
      type: object
      required:
        - firstname
        - lastname
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minimum: 6
          format: password
        firstname:
          type: string
          format: string
        lastname:
          type: string
          format: string

    Get User Response:
      type: object
      required:
        - uid
        - email
      properties:
        uid:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
          format: string
        photoUrl:
          type: string
          format: uri
        phoneNumber:
          type: string
          format: string
          pattern: "^[0-9]{10,15}$"
          example: +621234567891

    Edit User:
      type: object
      required:
        - uid
        - email
      properties:
        uid:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
          format: string
        image:
          type: string
          format: binary
        phoneNumber:
          type: string
          format: string
          pattern: "^[0-9]{10,15}$"
          example: +621234567891

    Edit User Response:
      type: object
      properties:
        message:
          type: string
          format: string
        data:
          type: object
          properties:
            uid:
              type: string
              format: uuid
            name:
              type: string
              format: string
            email:
              type: string
              format: email
            role:
              type: string
              enum: [user, partner, approver, admin]

    Reset Password:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email